fix: changed file structures
