+++ src/Components/Navigation/Navigation.scss
@@ -1,9 +1,3 @@
-.page-frame {
-  display: flex;
-  background-color: #e6eaec;
-  height: 100vh;
-  color: #fff;
-}
 .navigation {
   width: 20%;
   background-color: #327feb;
@@ -28,8 +22,4 @@
       text-decoration: underline;
     }
   }
-}
-
-.info-area {
-  width: 80%;
 }
