+++ src/Views/MainView/MainView.scss
@@ -1,0 +1,11 @@
+.page-frame {
+  display: flex;
+  background-color: #e6eaec;
+  height: 100vh;
+  color: #fff;
+}
+
+
+.info-area {
+  width: 80%;
+}
