+++ src/pages/LoginPage/LoginPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Login.scss';
+import './LoginPage.scss';
 import { Store } from 'react-notifications-component';
 import 'react-notifications-component/dist/theme.css';
 import 'animate.css';
